requirements.txt
slacker==0.12.0
PyYAML==4.2b1
tinydb==3.12.2
requests==2.20.0
pyOpenSSL==17.5.0
psutil==2.2.1
